refactor: 패키지 구조 이동 (#114)
