--- src/main/java/ku_rum/backend/domain/mail/dto/response/MailVerificationResponse.java
@@ -1,18 +1,0 @@
-package ku_rum.backend.domain.mail.dto.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class MailVerificationResponse {
-    private boolean verified;
-
-    private MailVerificationResponse(boolean verified) {
-        this.verified = verified;
-    }
-
-    public static MailVerificationResponse of(boolean verified) {
-        return new MailVerificationResponse(verified);
-    }
-}
