--- src/main/java/ku_rum/backend/global/handler/NoticeExceptionControllerAdvice.java
@@ -1,25 +1,0 @@
-package ku_rum.backend.global.handler;
-
-import jakarta.annotation.Priority;
-import ku_rum.backend.global.exception.notice.NoSuchNoticeException;
-import ku_rum.backend.global.response.BaseErrorResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import static ku_rum.backend.global.response.status.BaseExceptionResponseStatus.*;
-
-@Slf4j
-@Priority(0)
-@RestControllerAdvice
-public class NoticeExceptionControllerAdvice {
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(NoSuchNoticeException.class)
-    public BaseErrorResponse handleNoSuchNoticeException(final NoSuchNoticeException e) {
-        log.error("[NoSuchNoticeException]");
-        return new BaseErrorResponse(NO_SUCH_NOTICE);
-    }
-}
