--- src/main/java/ku_rum/backend/global/response/status/ResponseStatus.java
@@ -1,10 +1,0 @@
-package ku_rum.backend.global.response.status;
-
-import org.springframework.http.HttpStatus;
-
-public interface ResponseStatus {
-
-    int getCode();
-    HttpStatus getStatus();
-    String getMessage();
-}
