+++ src/main/java/ku_rum/backend/global/support/response/status/ResponseStatus.java
@@ -1,0 +1,10 @@
+package ku_rum.backend.global.support.response.status;
+
+import org.springframework.http.HttpStatus;
+
+public interface ResponseStatus {
+
+    int getCode();
+    HttpStatus getStatus();
+    String getMessage();
+}
